requirements.txt
# Core dependencies for Numismatch Agent
google-adk>=1.15.0,<2.0.0
google-cloud-aiplatform[evaluation,agent-engines]>=1.118.0,<2.0.0
google-cloud-logging>=3.12.0,<4.0.0
opentelemetry-exporter-gcp-trace>=1.9.0,<2.0.0
protobuf>=6.31.1,<7.0.0
openai>=1.0.0,<2.0.0  # Perplexity uses OpenAI-compatible client

# CLI for deployment
click>=8.1.0

# Optional: for local development and testing
# pytest>=8.3.4
# pytest-asyncio>=0.23.8

